wsl: tidy up gpg socket forwarding
`setsid` is sufficient. `nohup` is unnecessary because `setsid` is
running `socat` in a separate process, meaning it won't be HUPed or
killed based on the shell's lifecycle. `&` is both redundant (we don't
need to background because `setsid` is running `socat` in a separate
session) and incorrect (we're backgrounding `setsid`, oops).